#1 Adding resource-based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,13 +1,16 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes;
 
 public class DeleteDishesForRestaurantCommandHandler(IRestaurantsRepository restaurantsRepository,
 	IDishRepository dishRepository,
+	IRestaurantAuthorizationService restaurantAuthorizationService,
 	ILogger<DeleteDishesForRestaurantCommandHandler> logger) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
 	public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
@@ -16,10 +19,11 @@ public class DeleteDishesForRestaurantCommandHandler(IRestaurantsRepository rest
 
 		var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
 
-		if (restaurant == null)
-		{
+		if (restaurant == null)		
 			throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
-		}
+
+		if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+			throw new ForbidException();
 
 		await dishRepository.Delete(restaurant.Dishes);
 	}
