+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,12 +1,16 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
-public class DeleteRestaurantCommandHandler(IRestaurantsRepository restaurantsRepository,
+public class DeleteRestaurantCommandHandler(
+	IRestaurantsRepository restaurantsRepository,
+	IRestaurantAuthorizationService restaurantAuthorizationService,
 	ILogger<DeleteRestaurantCommandHandler> logger) : IRequestHandler<DeleteRestaurantCommand>
 {
 	public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -15,9 +19,10 @@ public class DeleteRestaurantCommandHandler(IRestaurantsRepository restaurantsRe
 		var restaurant = await restaurantsRepository.GetByIdAsync(request.Id);
 
 		if (restaurant is null)
-		{
 			throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
-		}
+
+		if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+			throw new ForbidException();
 
 		await restaurantsRepository.DeleteAsync(restaurant);
 	}
